requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.3
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
pillow==10.1.0
numpy==1.26.3
python-multipart==0.0.9
python-dotenv==1.0.0
mlflow==2.11.0
tensorflow==2.15.0
scikit-learn==1.3.2
matplotlib==3.8.2
pandas==2.1.4